BF: MPU stores/returns AWS-compat headers
Follow-up to PR #283, issue #282
Modify completeMultipartUpload so that Cache-Control, Content-Disposition, Content-Encoding, and Expires are stored in MD from shadow bucket for a get response. Additionally, refactor so that these headers are referenced more cleanly (by individual properties in the params object) instead of depending upon the whole request.headers object being passed as part of the metadataStoreParams / params object in metadataStoreObject.